re-api-daily-process/app/requirements.txt
pip-tools
psycopg2-binary==2.8.4
environ-config
google-auth==2.3.3
google-auth-oauthlib==0.4.6
pylint==2.4.4
pandas==1.3.0
requests==2.22.0
unidecode==1.1.1
pyathena
xlwt
SQLAlchemy==1.4.0
psutil==5.8.0
joblib==1.0.0
pandas-gbq==0.15.0
google-cloud-bigquery==2.28.0
pyarrow==5.0.0